# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DeciFlow API",
        "description": "API documentation for DeciFlow - Purchase Request Approval System",
        "contact": {
            "email": "support@deciflow.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/requests": {
            "get": {
                "tags": [
                    "Requests"
                ],
                "summary": "List all requests",
                "description": "Get paginated list of requests with optional filters",
                "operationId": "e8ba1b0f112480d4aad0b1360cb1a20a",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "DRAFT",
                                "SUBMITTED",
                                "IN_REVIEW",
                                "RETURNED",
                                "APPROVED",
                                "REJECTED",
                                "CANCELLED",
                                "ARCHIVED"
                            ]
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "EQUIPMENT",
                                "SOFTWARE",
                                "SERVICE",
                                "TRAVEL"
                            ]
                        }
                    },
                    {
                        "name": "department_id",
                        "in": "query",
                        "description": "Filter by department ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "amount_min",
                        "in": "query",
                        "description": "Minimum amount",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "amount_max",
                        "in": "query",
                        "description": "Maximum amount",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of requests",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Requests"
                ],
                "summary": "Create new request",
                "description": "Create a new purchase request in DRAFT status",
                "operationId": "927dd34168a9440882d2d9d4aea3091e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "category",
                                    "amount"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "New laptop for development"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "MacBook Pro 16-inch for development work"
                                    },
                                    "category": {
                                        "type": "string",
                                        "enum": [
                                            "EQUIPMENT",
                                            "SOFTWARE",
                                            "SERVICE",
                                            "TRAVEL"
                                        ],
                                        "example": "EQUIPMENT"
                                    },
                                    "amount": {
                                        "description": "Amount in JPY",
                                        "type": "integer",
                                        "example": 300000
                                    },
                                    "vendor_name": {
                                        "type": "string",
                                        "example": "Apple Inc."
                                    },
                                    "urgency": {
                                        "type": "string",
                                        "enum": [
                                            "NORMAL",
                                            "URGENT"
                                        ],
                                        "example": "NORMAL"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Request created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/requests/{id}": {
            "get": {
                "tags": [
                    "Requests"
                ],
                "summary": "Get request details",
                "description": "Get detailed information about a specific request",
                "operationId": "ccb3309007cab7ed5d86ecc7c17cb037",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Request not found"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Requests"
                ],
                "summary": "Update request",
                "description": "Update a request in DRAFT status",
                "operationId": "4f35861736adb03021ffaef6b8bf458d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string",
                                        "enum": [
                                            "EQUIPMENT",
                                            "SOFTWARE",
                                            "SERVICE",
                                            "TRAVEL"
                                        ]
                                    },
                                    "amount": {
                                        "type": "integer"
                                    },
                                    "vendor_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Cannot update non-DRAFT request"
                    },
                    "404": {
                        "description": "Request not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/requests/{id}/submit": {
            "post": {
                "tags": [
                    "Requests"
                ],
                "summary": "Submit request",
                "description": "Submit a request for approval",
                "operationId": "aecd6537f25a2ae9f61df2cda92dbcbd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Submit failed"
                    },
                    "404": {
                        "description": "Request not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/requests/{id}/cancel": {
            "post": {
                "tags": [
                    "Requests"
                ],
                "summary": "Cancel request",
                "description": "Cancel a submitted request",
                "operationId": "8e0fd0ab3d0d9798e65ca31181f6340f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request cancelled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Cancel failed"
                    },
                    "404": {
                        "description": "Request not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/requests/{id}/archive": {
            "post": {
                "tags": [
                    "Requests"
                ],
                "summary": "Archive request",
                "description": "Archive a request (admin only)",
                "operationId": "e0ae9da2fd4de605401b06836892a9b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Request ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request archived successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - admin only"
                    },
                    "422": {
                        "description": "Archive failed"
                    },
                    "404": {
                        "description": "Request not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}